requirements.txt
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
Django==2.0.13
django-extra-views==0.11.0
django-haystack==2.8.1
django-localflavor==2.2
django-oscar==2.0.2
django-oscar-paypal==1.0.0
django-phonenumber-field==2.0.1
django-tables2==1.21.2
django-treebeard==4.3
django-widget-tweaks==1.4.5
factory-boy==2.12.0
Faker==2.0.2
idna==2.8
phonenumbers==8.10.20
Pillow==6.2.0
purl==1.5
pycountry==19.8.18
python-dateutil==2.8.0
python-stdnum==1.11
pytz==2019.2
requests==2.22.0
six==1.12.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.6
